+++ src/textual_pyfiglet/demo.py
@@ -3,8 +3,9 @@ This module contains the demo application for PyFiglet.
 
 It has its own entry script. Run with `textual-pyfiglet`.
 """
+
 # Python imports
-from typing import cast
+from typing import Any  # , cast
 from importlib import resources
 import re
 import random
@@ -18,22 +19,23 @@ from textual.binding import Binding
 from textual.screen import ModalScreen
 from textual.color import Color, ColorParseError
 from textual.validation import Validator, ValidationResult, Number
-from textual.widgets import (
-    Header, Footer, Static, Input, TextArea,
-    Select, Switch, Label, Markdown, Button
-)
+from textual.widgets import Header, Footer, Static, Input, TextArea, Select, Switch, Label, Markdown, Button
+
+# from textual.widgets._select import NoSelection
 
 # textual-pyfiglet imports
-from textual_pyfiglet.figletwidget import FigletWidget, JUSTIFY_OPTIONS
-from textual_pyfiglet.pyfiglet.fonts import ALL_FONTS
+from textual_pyfiglet.figletwidget import FigletWidget  # , JUSTIFY_OPTIONS
+
+# from textual_pyfiglet.pyfiglet.fonts import ALL_FONTS
 from textual_slidecontainer import SlideContainer
 
 from rich import traceback
 from rich.text import Text
+
 traceback.install()
 
 
-class HelpScreen(ModalScreen):
+class HelpScreen(ModalScreen[None]):
 
     BINDINGS = [
         Binding("escape,enter", "close_screen", description="Close the help window.", show=True),
@@ -41,10 +43,10 @@ class HelpScreen(ModalScreen):
 
     def compose(self):
 
-        with resources.open_text('textual_pyfiglet', 'help.md') as f:
+        with resources.open_text("textual_pyfiglet", "help.md") as f:
             self.help = f.read()
 
-        with VerticalScroll(id='help_container'):
+        with VerticalScroll(id="help_container"):
             yield Markdown(self.help)
 
     def on_mount(self):
@@ -56,6 +58,7 @@ class HelpScreen(ModalScreen):
     def action_close_screen(self):
         self.dismiss()
 
+
 class SettingBox(Container):
 
     def __init__(
@@ -64,13 +67,11 @@ class SettingBox(Container):
         label: str = "",
         label_position: str = "beside",
         widget_width: int | None = None,
-        *args,
-        **kwargs
     ):
         """A setting box with a label and a widget. \n
         Label position can be either 'beside' or 'under'"""
 
-        super().__init__(*args, **kwargs)
+        super().__init__()
         self.widget = widget
         self.label = label
         self.label_position = label_position
@@ -86,7 +87,7 @@ class SettingBox(Container):
                 yield Static(classes="setting_filler")
                 if self.label:
                     yield Label(self.label, classes="setting_label")
-                yield self.widget                
+                yield self.widget
         elif self.label_position == "under":
             with Horizontal():
                 yield Static(classes="setting_filler")
@@ -101,45 +102,44 @@ class SettingBox(Container):
 class SizeValidator(Validator):
 
     patterns = [
-        r'^[1-9][0-9]{0,2}$',       # Number between 1-999
-        r'^(100|[1-9]?[0-9])%$',    # Percentage
-        r'^\d*\.?\d+fr$',           # Float followed by 'fr'
+        r"^[1-9][0-9]{0,2}$",  # Number between 1-999
+        r"^(100|[1-9]?[0-9])%$",  # Percentage
+        r"^\d*\.?\d+fr$",  # Float followed by 'fr'
     ]
 
     def validate(self, value: str) -> ValidationResult:
 
         if any(re.match(pattern, value) for pattern in self.patterns):
             return self.success()
-        elif value == '':
+        elif value == "":
             return self.success()
         else:
             return self.failure(
                 "Invalid size format. Must be a number between 1-999, a percentage, "
                 "a float followed by 'fr', or 'auto'."
             )
-        
+
 
 class ColorValidator(Validator):
 
     def validate(self, value: str) -> ValidationResult:
 
-        if value == '':
+        if value == "":
             return self.success()
 
         try:
             Color.parse(value)
         except ColorParseError:
-            return self.failure(
-                "Invalid color format. Must be a valid color name or hex code."
-            )
+            return self.failure("Invalid color format. Must be a valid color name or hex code.")
         else:
             return self.success()
-        
+
+
 class QualityValidator(Validator):
 
     def validate(self, value: str) -> ValidationResult:
 
-        if value == '':
+        if value == "":
             return self.success()
         try:
             value_int = int(value)
@@ -154,54 +154,52 @@ class QualityValidator(Validator):
                 return self.failure(
                     "Invalid quality format. Must be empty (for auto), or an integer between 3-100."
                 )
-          
+
 
 class SettingsWidget(VerticalScroll):
 
     justifications = [
-        ('Left', 'left'),
-        ('Center', 'center'),
-        ('Right', 'right'),
-    ]  
-
-    patterns = [
-        r'^[1-9][0-9]{0,2}$',       # Number between 1-999
-        r'^(100|[1-9]?[0-9])%$',    # Percentage
-        r'^\d*\.?\d+fr$',           # Float followed by 'fr'
+        ("Left", "left"),
+        ("Center", "center"),
+        ("Right", "right"),
     ]
 
-    def __init__(self, figlet_widget: FigletWidget, *args, **kwargs):
-        super().__init__(*args, **kwargs)
+    patterns = [
+        r"^[1-9][0-9]{0,2}$",  # Number between 1-999
+        r"^(100|[1-9]?[0-9])%$",  # Percentage
+        r"^\d*\.?\d+fr$",  # Float followed by 'fr'
+    ]
+
+    def __init__(self, figlet_widget: FigletWidget):
+        super().__init__()
         self.figlet_widget = figlet_widget
+        self.fonts_list = self.figlet_widget.fonts_list
+        self.fonts_list.sort()
+        self.font_options = [(font, font) for font in self.fonts_list]
 
     def compose(self):
 
-        self.fonts_list = self.figlet_widget.get_fonts_list()
-        self.fonts_list.sort()
-        self.font_options = [(font, font) for font in self.fonts_list] 
-
         self.randomize = Button("Random Font", id="randomize_button")
-        self.font_select = Select(
-            self.font_options, value="ansi_regular", id="font_select", allow_blank=True
-        )
-        self.width_input  = Input(id="width_input", validators=[SizeValidator()], max_length=5)
+        self.font_select = Select(self.font_options, value="ansi_regular", id="font_select", allow_blank=True)
+        self.width_input = Input(id="width_input", validators=[SizeValidator()], max_length=5)
         self.height_input = Input(id="height_input", validators=[SizeValidator()], max_length=5)
         self.justify_select = Select(
-            self.justifications, value='center', id="justify_select", allow_blank=False
-            )
-        self.padding_input = Input(
-            value="0", id="padding_input", validators=[Number()], max_length=2
+            self.justifications, value="center", id="justify_select", allow_blank=False
         )
+        self.padding_input = Input(value="0", id="padding_input", validators=[Number()], max_length=2)
         self.color1_input = Input(id="color1_input", validators=[ColorValidator()])
         self.color2_input = Input(id="color2_input", validators=[ColorValidator()])
         self.animate_switch = Switch(id="animate_switch", value=False)
         self.gradient_quality = Input(
-            id="gradient_quality", max_length=3, 
-            validators=[QualityValidator()], 
+            id="gradient_quality",
+            max_length=3,
+            validators=[QualityValidator()],
         )
         self.animation_speed = Input(
-            id="animation_speed", value="0.08", max_length=4,
-            validators=[Number(minimum=0.05, maximum=1.0)], 
+            id="animation_speed",
+            value="0.08",
+            max_length=4,
+            validators=[Number(minimum=0.01, maximum=1.0)],
         )
 
         yield Label("Settings", id="settings_title")
@@ -224,31 +222,30 @@ class SettingsWidget(VerticalScroll):
 
         self.log("Randomizing font...")
 
-        fonts = self.figlet_widget.get_fonts_list()
+        fonts = self.figlet_widget.fonts_list
         self.font_select.value = random.choice(fonts)
 
-
-    @on(Select.Changed, selector="#font_select")           
+    @on(Select.Changed, selector="#font_select")
     def font_changed(self, event: Select.Changed) -> None:
 
-        if event.value == Select.BLANK:   #! Explain why blank is even allowed.
+        if event.value == Select.BLANK:  #! Explain why blank is even allowed.
             return
-        
-        self.log(f"Setting font to: {event.value}...")         
-        self.figlet_widget.set_font(cast(ALL_FONTS, event.value))
+
+        self.log(f"Setting font to: {event.value}...")
+        self.figlet_widget.font = str(event.value)
 
     @on(Input.Submitted, selector="#width_input")
     @on(Input.Blurred, selector="#width_input")
     def width_input_set(self, event: Input.Blurred) -> None:
 
         if event.validation_result:
-            if event.validation_result.is_valid:          
+            if event.validation_result.is_valid:
                 self.log(f"Width set to: {event.value}")
-                width = self.width_input.value if self.width_input.value != '' else 'auto'
-                height = self.height_input.value if self.height_input.value != '' else 'auto'
+                width = self.width_input.value if self.width_input.value != "" else "auto"
+                height = self.height_input.value if self.height_input.value != "" else "auto"
                 self.log(f"Setting container size to: ({width} x {height})")
 
-                if width == 'auto':
+                if width == "auto":
                     self.figlet_widget.styles.width = width
                     self.log(f"Width set to: {self.figlet_widget.styles.width}")
                 else:
@@ -257,23 +254,23 @@ class SettingsWidget(VerticalScroll):
                         self.log(f"Width set to integer: {self.figlet_widget.styles.width}")
                     except ValueError:
                         self.figlet_widget.styles.width = width
-                        self.log(f"Width set to: {self.figlet_widget.styles.width}")               
-            else: 
-                failures = event.validation_result.failure_descriptions   
+                        self.log(f"Width set to: {self.figlet_widget.styles.width}")
+            else:
+                failures = event.validation_result.failure_descriptions
                 self.log(f"Invalid width: {failures}")
-                self.notify(f"Invalid width: {failures}", markup=False)    
+                self.notify(f"Invalid width: {failures}", markup=False)
 
     @on(Input.Submitted, selector="#height_input")
     @on(Input.Blurred, selector="#height_input")
     def height_input_set(self, event: Input.Blurred) -> None:
 
         if event.validation_result:
-            if event.validation_result.is_valid:          
-                width = self.width_input.value if self.width_input.value != '' else 'auto'
-                height = self.height_input.value if self.height_input.value != '' else 'auto'
+            if event.validation_result.is_valid:
+                width = self.width_input.value if self.width_input.value != "" else "auto"
+                height = self.height_input.value if self.height_input.value != "" else "auto"
                 self.log(f"Setting container size to: ({width} x {height})")
-                
-                if height == 'auto':
+
+                if height == "auto":
                     self.figlet_widget.styles.height = height
                     self.log(f"Height set to: {self.figlet_widget.styles.height}")
                 else:
@@ -283,27 +280,27 @@ class SettingsWidget(VerticalScroll):
                     except ValueError:
                         self.figlet_widget.styles.height = height
                         self.log(f"Height set to: {self.figlet_widget.styles.height}")
-            else: 
-                failures = event.validation_result.failure_descriptions  
+            else:
+                failures = event.validation_result.failure_descriptions
                 self.log(f"Invalid height: {failures}")
                 self.notify(f"Invalid height: {failures}", markup=False)
 
     @on(Select.Changed, selector="#justify_select")
     def justify_changed(self, event: Select.Changed) -> None:
-        
+
         self.log(f"Setting justify to: {event.value}...")
-        self.figlet_widget.set_justify(cast(JUSTIFY_OPTIONS, event.value))
-   
+        self.figlet_widget.justify = str(event.value)
+
     @on(Input.Submitted, selector="#padding_input")
     @on(Input.Blurred, selector="#padding_input")
     def padding_input_set(self, event: Input.Blurred) -> None:
 
         if event.validation_result:
-            if event.validation_result.is_valid:                            
+            if event.validation_result.is_valid:
                 self.log(f"Padding set to: {event.value}")
                 self.figlet_widget.styles.padding = int(event.value)
-            else: 
-                failures = event.validation_result.failure_descriptions   
+            else:
+                failures = event.validation_result.failure_descriptions
                 self.log(f"Invalid padding input: {failures}")
 
     @on(Input.Submitted, selector="#color1_input")
@@ -311,12 +308,12 @@ class SettingsWidget(VerticalScroll):
     def color1_input_set(self, event: Input.Blurred) -> None:
 
         if event.validation_result:
-            if event.validation_result.is_valid:                        
+            if event.validation_result.is_valid:
                 self.log(f"Color1 set to: {event.value}")
-                self.figlet_widget.set_color1(event.value)
-            else: 
-                failures = event.validation_result.failure_descriptions 
-                self.log(f"Invalid color1 input: {failures}")  
+                self.figlet_widget.color1 = event.value
+            else:
+                failures = event.validation_result.failure_descriptions
+                self.log(f"Invalid color1 input: {failures}")
                 self.notify(f"Invalid color1 input: {failures}", markup=False)
 
     @on(Input.Submitted, selector="#color2_input")
@@ -324,18 +321,18 @@ class SettingsWidget(VerticalScroll):
     def color2_input_set(self, event: Input.Blurred) -> None:
 
         if event.validation_result:
-            if event.validation_result.is_valid:                        
+            if event.validation_result.is_valid:
                 self.log(f"Color2 set to: {event.value}")
-                self.figlet_widget.set_color2(event.value)
-            else: 
-                failures = event.validation_result.failure_descriptions 
-                self.log(f"Invalid color2 input: {failures}") 
-                self.notify(f"Invalid color2 input: {failures}", markup=False)   
+                self.figlet_widget.color2 = event.value
+            else:
+                failures = event.validation_result.failure_descriptions
+                self.log(f"Invalid color2 input: {failures}")
+                self.notify(f"Invalid color2 input: {failures}", markup=False)
 
-    @on(Switch.Changed, selector="#animate_switch")    
+    @on(Switch.Changed, selector="#animate_switch")
     def animate_switch_toggled(self, event: Switch.Changed) -> None:
 
-        self.figlet_widget.set_animated(event.value)
+        self.figlet_widget.animated = event.value
 
     @on(Input.Submitted, selector="#gradient_quality")
     @on(Input.Blurred, selector="#gradient_quality")
@@ -345,14 +342,14 @@ class SettingsWidget(VerticalScroll):
         Auto mode will set the quality to the height of the widget."""
 
         if event.validation_result:
-            if event.validation_result.is_valid:                            
+            if event.validation_result.is_valid:
                 self.log(f"Gradient quality set to: {event.value}")
-                if event.value == '':
-                    self.figlet_widget.set_gradient_quality('auto')
+                if event.value == "":
+                    self.figlet_widget.gradient_quality = "auto"
                 else:
-                    self.figlet_widget.set_gradient_quality(int(event.value))
-            else: 
-                failures = event.validation_result.failure_descriptions   
+                    self.figlet_widget.gradient_quality = int(event.value)
+            else:
+                failures = event.validation_result.failure_descriptions
                 self.log(f"Invalid Gradient quality input: {failures}")
                 self.notify(f"Invalid Gradient quality input: {failures}", markup=False)
 
@@ -363,20 +360,20 @@ class SettingsWidget(VerticalScroll):
         This must be a number between 0.05 - 1.0"""
 
         if event.validation_result:
-            if event.validation_result.is_valid:                            
+            if event.validation_result.is_valid:
                 self.log(f"Animation speed set to: {event.value}")
-                self.figlet_widget.set_animation_speed(float(event.value))
-            else: 
-                failures = event.validation_result.failure_descriptions   
+                self.figlet_widget.animation_interval = float(event.value)
+            else:
+                failures = event.validation_result.failure_descriptions
                 self.log(f"Invalid animation speed input: {failures}")
                 self.notify(f"Invalid animation speed input: {failures}", markup=False)
 
 
 class BottomBar(Horizontal):
 
-    def __init__(self, figlet_widget: FigletWidget, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.figlet_widget = figlet_widget   
+    def __init__(self, figlet_widget: FigletWidget):
+        super().__init__()
+        self.figlet_widget = figlet_widget
 
     def compose(self):
 
@@ -390,17 +387,18 @@ class BottomBar(Horizontal):
 
         # This just scrolls the text area to the end when the text changes:
         scroll_area = self.app.query_one("#main_window")
-        if scroll_area.scrollbars_enabled == (True, False): # Vertical True, Horizontal False
+        if scroll_area.scrollbars_enabled == (True, False):  # Vertical True, Horizontal False
             scroll_area.action_scroll_end()
 
     def focus_textarea(self):
         # Used when the demo boots to focus the text input.
-        
+
         self.text_input.focus()
         end = self.text_input.get_cursor_line_end_location()
         self.text_input.move_cursor(end)
 
-class TextualPyFigletDemo(App):
+
+class TextualPyFigletDemo(App[Any]):
 
     #! Needs better bindings.
     BINDINGS = [
@@ -411,27 +409,25 @@ class TextualPyFigletDemo(App):
     CSS_PATH = "styles.tcss"
     TITLE = "Textual-PyFiglet Demo"
 
-    def on_resize(self):                    
-        self.figlet_widget.refresh_size()   # <-- This is how you make it resize automatically.
+    def on_resize(self):
+        self.figlet_widget.refresh_size()  # <-- This is how you make it resize automatically.
 
     def compose(self):
 
-        self.figlet_widget = FigletWidget(      #~ <--- This is the main widget.
-            "Starter Text",                     # You can input all kinds of arguments directly.
-            id="figlet_widget",                 # But for the purposes of the demo, all these
-            # color1="red",                 # Every setting can be changed in real-time.
-            # color2="blue",          
+        self.figlet_widget = FigletWidget(  # ~ <--- This is the main widget.
+            "Starter Text",  # You can input all kinds of arguments directly.
+            id="figlet_widget",  # But for the purposes of the demo, all these
+            # color1="red",      # are set in real-time in the demo sidebar.
+            # color2="blue",
         )
 
         banner = FigletWidget.figlet_quick("Textual-PyFiglet", font="smblock")
         self.log(Text.from_markup(f"[bold blue]{banner}"))
 
-        self.settings_widget  = SettingsWidget(self.figlet_widget)
-        self.bottom_bar       = BottomBar(self.figlet_widget)
+        self.settings_widget = SettingsWidget(self.figlet_widget)
+        self.bottom_bar = BottomBar(self.figlet_widget)
         self.size_display_bar = Static(id="size_display", expand=True)
-        self.menu_container   = SlideContainer(
-            id="menu_container", slide_direction="left", floating=False
-        )
+        self.menu_container = SlideContainer(id="menu_container", slide_direction="left", floating=False)
 
         # Note: Layout is horizontal. (top of styles.tcss)
         yield Header(name="Textual-PyFiglet Demo")
@@ -467,7 +463,8 @@ class TextualPyFigletDemo(App):
     def action_show_help(self):
         self.push_screen(HelpScreen())
 
+
 # This is for the entry script. Run the demo with:
-#$ textual-pyfiglet
+# $ textual-pyfiglet
 def run_demo():
     TextualPyFigletDemo().run()